move comparators to image_tools
