+++ packages/playwright-core/package.json
@@ -27,9 +27,9 @@
     "./lib/image_tools/compare": "./lib/image_tools/compare.js",
     "./lib/image_tools/imageChannel": "./lib/image_tools/imageChannel.js",
     "./lib/image_tools/colorUtils": "./lib/image_tools/colorUtils.js",
+    "./lib/image_tools/comparators": "./lib/image_tools/comparators.js",
     "./lib/common/userAgent": "./lib/common/userAgent.js",
     "./lib/containers/docker": "./lib/containers/docker.js",
-    "./lib/utils/comparators": "./lib/utils/comparators.js",
     "./lib/utils/eventsHelper": "./lib/utils/eventsHelper.js",
     "./lib/utils/fileUtils": "./lib/utils/fileUtils.js",
     "./lib/utils/httpServer": "./lib/utils/httpServer.js",
