+++ packages/playwright-core/src/image_tools/comparators.ts
