+++ packages/playwright-core/src/server/page.ts
@@ -34,8 +34,8 @@ import { ProgressController } from './progress';
 import { assert, isError } from '../utils';
 import { ManualPromise } from '../utils/manualPromise';
 import { debugLogger } from '../common/debugLogger';
-import type { ImageComparatorOptions } from '../utils/comparators';
-import { getComparator } from '../utils/comparators';
+import type { ImageComparatorOptions } from '../image_tools/comparators';
+import { getComparator } from '../image_tools/comparators';
 import type { SelectorInfo, Selectors } from './selectors';
 import type { CallMetadata } from './instrumentation';
 import { SdkObject } from './instrumentation';
